refactor: InputValidator Menu로 분리
- 테스트 코드도 분리
