+++ src/test/java/com/devcourse/java/calculator/ConsoleTest.java
@@ -1,6 +1,7 @@
 package com.devcourse.java.calculator;
 
 import com.devcourse.java.calculator.constant.ExceptionMessageConstant;
+import com.devcourse.java.calculator.constant.Menu;
 import com.devcourse.java.calculator.io.Console;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,40 +18,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class ConsoleTest {
 
     Console console = new Console();
-
-    @Test
-    @DisplayName("메뉴 선택 입력이 정수가 아닐 경우 InputMismatchException 확인")
-    void getCommand_Input_Not_Integer() {
-        //given
-        String command1 = "abc";
-        String command2 = ".~!,=";
-
-        //when, then
-        assertThatThrownBy(() -> InputValidator.checkCommandInput(command1))
-                .isInstanceOf(InputMismatchException.class)
-                .hasMessageContaining(ExceptionMessageConstant.COMMAND_INPUT_NOT_INTEGER_EXCEPTION);
-
-        assertThatThrownBy(() -> InputValidator.checkCommandInput(command2))
-                .isInstanceOf(InputMismatchException.class)
-                .hasMessageContaining(ExceptionMessageConstant.COMMAND_INPUT_NOT_INTEGER_EXCEPTION);
-    }
-
-    @Test
-    @DisplayName("메뉴 선택 입력이 정수지만 1~3이 아닐 경우 InputMismatchException 확인")
-    void getCommand_Input_Not_In_Boundary() {
-        //given
-        String command1 = "99";
-        String command2 = "30";
-
-        //when, then
-        assertThatThrownBy(() -> InputValidator.checkCommandInput(command1))
-                .isInstanceOf(InputMismatchException.class)
-                .hasMessageContaining(ExceptionMessageConstant.COMMAND_INPUT_NOT_IN_BOUNDARY_EXCEPTION);
-
-        assertThatThrownBy(() -> InputValidator.checkCommandInput(command2))
-                .isInstanceOf(InputMismatchException.class)
-                .hasMessageContaining(ExceptionMessageConstant.COMMAND_INPUT_NOT_IN_BOUNDARY_EXCEPTION);
-    }
 
     @Test
     @DisplayName("비어있는 계산 내역 조회할때 IllegalArgumentException 확인")
